\name{landsnail}
\alias{landsnail}
\docType{data}
\title{Species Attributes for 121 Hungarian Land Snails}
\description{A data frame with species attributes for 121 Hungarian land snail species.

}
\usage{data(landsnail)}
\format{
A data frame with eight attributes of 121 land snail species. Attributes are: ranks (for taxonomic ordering), species names (short and full), authors plus description dates, adult body size and higher level taxonomy (familia, subfamilia, genus).
}
\source{
Kerney, M.P., Cameron, R.A.D. & Jungbluth, J.H., 1983. Die Landschnecken Nord- und Mitteleuropas. P. Parey, Hamburg-Berlin, 384 p.
}
\examples{
data(landsnail)
}
\keyword{datasets}
