\name{mefaweb}
\alias{mefaweb}
\title{Opens the 'mefa' website in the default navigator}
\description{
Opens the 'mefa' website in the default navigator
}
\usage{
mefaweb()
}
\details{
The website contains patches and beta releases of the'mefa' package, links to related documents and case studies.
}
\author{Peter Solymos, \email{Solymos.Peter@aotk.szie.hu}, \url{http://www.univet.hu/users/psolymos/personal/}}
\examples{
\dontrun{
mefaweb()
}
}
\keyword{ utilities }

