\name{mefaweb}
\alias{mefaweb}
\title{Opens the 'mefa' website}
\description{
Opens the 'mefa' website in the default navigator.
}
\usage{
mefaweb()
}
\details{
The website contains resources for the 'mefa' package, and many more. Please check out!
}
\author{Peter Solymos, \email{Solymos.Peter@aotk.szie.hu}, \url{http://www.univet.hu/users/psolymos/personal/}}
\examples{
\dontrun{
mefaweb()
}
}
\keyword{ utilities }
