\name{mefaweb}
\alias{mefaweb}
\title{Open 'mefa' website}
\description{
Opens the 'mefa' website in the default navigator.
}
\usage{
mefaweb()
}
\details{
The website contains resources for the 'mefa' package, please check out for more. Thank you!
}
\author{
Peter Solymos, \email{solymos@ualberta.ca}
}

\examples{
\dontrun{
mefaweb()
}
}
\keyword{ documentation }
