\name{mefaweb}
\alias{mefaweb}
\title{Open 'mefa' Website}
\description{
Opens the 'mefa' website in the default navigator.
}
\usage{
mefaweb()
}
\details{
The website contains resources for the 'mefa' package, and many more. Please check out for more!
}
\author{
Peter Solymos, \email{solymos@ualberta.ca}
}

\examples{
\dontrun{
mefaweb()
}
}
\keyword{ utilities }
