\encoding{UTF-8}
\name{mefaweb}
\alias{mefaweb}
\title{Open 'mefa' Website}
\description{
Opens the 'mefa' website in the default navigator.
}
\usage{
mefaweb()
}
\details{
The website contains resources for the 'mefa' package, please check out for more. Thank you!
}
\references{
\enc{Sólymos}{Solymos} P. 2008. mefa: an R package for handling and reporting count data. \emph{Community Ecology} 9, 125--127.

\url{http://mefa.r-forge.r-project.org/}
}
\author{
\enc{Péter Sólymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\examples{
\dontrun{
mefaweb()
}
}
\keyword{ documentation }
