\encoding{UTF-8}
\name{mefaweb}
\alias{mefaweb}
\title{Open 'mefa' Website}
\description{
Opens the 'mefa' website in the default navigator.
}
\usage{
mefaweb()
}
\details{
The website contains resources for the 'mefa' package, please check out for more. Thank you!
}
\references{
\enc{S\'{o}lymos}{Solymos} P. (2008) mefa: an R package for handling and reporting count data. \emph{Community Ecology} \bold{9}, 125--127.

\enc{S\'{o}lymos}{Solymos} P. (2009) Processing ecological data in R with the mefa package. \emph{Journal of Statistical Software} \bold{29(8)}, 1--28. \url{http://www.jstatsoft.org/v29/i08/}

\url{http://mefa.r-forge.r-project.org/}
}
\author{
\enc{P\'{e}ter S\'{o}lymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\examples{
\dontrun{
mefaweb()
}
}
\keyword{ documentation }
